+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added documentation on `Tornium.Telemetry` events
+
 ### Fixed
 - Fixed startup order of `Tornex.Scheduler.QueryRegistry` in `Tornex.Scheduler.Supervisor`
 - Fixed `Tornex.Scheduler.Timer` failing to dump queries in other nodes